Move remaining errors to common errors.go
Rule engine and repair manager errors not included, as they are in a different CR

Some effective duplicate errors removed.
