+++ internal/services/inventory/pdu_test.go
@@ -63,7 +63,7 @@ func (ts *PduTestSuite) TestBadPowerTarget() {
 	require.NotNil(res)
 
 	assert.Error(res.Err)
-	assert.Equal(messages.ErrInvalidTarget, res.Err)
+	assert.Equal(errors.ErrInvalidTarget, res.Err)
 	assert.Equal(common.TickFromContext(ctx), res.At)
 	assert.Nil(res.Msg)
 
@@ -338,7 +338,7 @@ func (ts *PduTestSuite) TestPowerOnBladeBadID() {
 	require.True(ok)
 	require.NotNil(res)
 	assert.Error(res.Err)
-	assert.Equal(messages.ErrInvalidTarget, res.Err)
+	assert.Equal(errors.ErrInvalidTarget, res.Err)
 
 	assert.Equal(common.TickFromContext(ctx), res.At)
 	assert.Less(r.pdu.sm.Guard, msg.Guard)
