+++ pkg/protos/common/timestamp.Validate.go
@@ -2,12 +2,16 @@
 
 package common
 
+import (
+    "github.com/Jim3Things/CloudChamber/pkg/errors"
+)
+
 // Validate is a method that verifies that the associated Timestamp instance
 // is structurally legal
 func (x *Timestamp) Validate() error {
 	// No negative time allowed
 	if x.Ticks < 0 {
-		return ErrMustBeGTE{Field: "Ticks", Actual: x.Ticks, Required: 0}
+		return errors.ErrMustBeGTE{Field: "Ticks", Actual: x.Ticks, Required: 0}
 	}
 
 	// All correct
